fix router on ipad
