+++ Count my breaths/App.swift
@@ -8,15 +8,43 @@
 
 import SwiftUI
 import ToastUI
+import StoreKit
 import QuickComponents
 
 struct App: View {
     
     @State var showNewVersion = false
     @State var version = VersionController.loadVersion()
+    @State var selected: Routes?
+    
+    init() {
+        let userDefaults = UserDefaults.standard
+        let appRuns = userDefaults.integer(forKey: "appruns")
+        let countRuns = userDefaults.integer(forKey: "countruns")
+        let reviewShown = userDefaults.bool(forKey: "reviewshown")
+        if (appRuns >= 5 || countRuns >= 5) && !reviewShown {
+            SKStoreReviewController.requestReview()
+            userDefaults.set(true, forKey: "reviewshown")
+        }
+        UITableView.appearance().backgroundColor = UIColor(red: 78/255, green: 78/255, blue: 78/255, alpha: 0.2)
+    }
     
     var body: some View {
-        Navigation()
+        ZStack {
+            Theme.colors.text.edgesIgnoringSafeArea(.all)
+            NavigationView {
+                Router(selected: $selected)
+                    .navigationBarTitle("Menu")
+                
+                CountView()
+                    .padding()
+                    .navigationBarTitle("Counter")
+            }
+        }.onAppear {
+            if UIDevice.current.userInterfaceIdiom == .phone {
+                self.selected = .counter
+            }
+        }
             .toast(isPresented: $showNewVersion) {
             ToastView {
                 VStack {
